feat: parse order of day/month/year based on value and `preferedOrder` option
`parseFormat` tries to figure out the the order of day/month/year by itself
if it finds 3 numbers separated by `.`, `-` or `/`. But if it can't, it will fallback
to `preferedOrder`, which can either be set as an object to differentiate by separator,
or as a simple string.

Default value:

```js
preferedOrder: {
  '/': 'MDY',
  '.': 'DMY',
  '-': 'YMD'
}
```

Usage

```js
parseFormat('10.10.2010', {preferedOrder: 'DMY'});
// ☛ DD.MM.YYYY
parseFormat('10.10.2010', {preferedOrder: 'MDY'});
// ☛ MM.DD.YYYY
parseFormat('10.10.2010', {preferedOrder: {
  '/': 'MDY',
  '.': 'DMY',
  '-': 'YMD'
}});
// ☛ MM.DD.YYYY
parseFormat('10/10/2010', {preferedOrder: {
  '/': 'MDY',
  '.': 'DMY',
  '-': 'YMD'
}});
// ☛ DD/MM/YYYY
```

fixes #11