Added support for all sort of odd formats
Support various other formats that are usually not standard, such as
those that mix D with MM, or M with DD, and many others that were not
properly detected before.

Also, improved timezone detection. Instead of the wild card approach
which would incorrectly classify years as time zones in certain cases, a
regex that matched only known timezone values was added.